% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam.R
\name{knit_print.tutorial_question}
\alias{knit_print.tutorial_question}
\title{Knitr question print methods}
\usage{
\method{knit_print}{tutorial_question}(x, ...)
}
\arguments{
\item{x}{An R object to be printed}

\item{...}{Additional arguments passed to the S3 method. Currently ignored,
except two optional arguments \code{options} and \code{inline}; see
the references below.}
}
\description{
\code{knitr::\link[knitr]{knit_print}}
}
